Move to atomic
